add ifdefs, move component declarations to .h files
... and don't use to_string with pointer in hashes, as this could easily change
